// src/model/match.ts
export type Match = {
    adj_score1: string;
    adj_score2: string;
    date: string;
    importance1: string;
    importance2: string;
    league: string;
    league_id: string;
    nsxg1: string;
    nsxg2: string;
    prob1: string;

    prob2: string;
    probtie: string;
    proj_score1: string;
    proj_score2: string;
    score1: string;
    score2: string;
    season: string;
    spi1: string;
    spi2: string;
    team1: string;
    team2: string;
    xg1: string;
    xg2: string;
};